+++ dflib-echarts/src/main/java/org/dflib/echarts/BoxplotSeriesOpts.java
@@ -1,0 +1,12 @@
+package org.dflib.echarts;
+
+/**
+ * @since 1.0.0-RC1
+ */
+public class BoxplotSeriesOpts extends CartesianSeriesOpts<BoxplotSeriesOpts> {
+
+    @Override
+    public ChartType getType() {
+        return ChartType.boxplot;
+    }
+}